basic MM to go straight

+++ src/main/kotlin/frc/team5190/robot/auto/AutoHelper.kt
@@ -279,5 +279,5 @@ data class TrajectoryData(private val position: Double, private val velocity: Do
 
     // Converts rotations and rotations/sec to native units and native units/100 ms.
     var nativeUnits = Maths.rotationsToNativeUnits(rotations, Hardware.NATIVE_UNITS_PER_ROTATION.toDouble())
-    val nativeUnitsPer100Ms = Maths.rpmToNativeUnitsPer100Ms(rpm, Hardware.NATIVE_UNITS_PER_ROTATION.toDouble())
+    val nativeUnitsPer100Ms = Maths.rpmToNativeUnitsPer100Ms(rpm, Hardware.NATIVE_UNITS_PER_ROTATION)
 }
